fix: Opamp bugfix (#33)
* feat: Add AgentCapabilities & minor bugfix

* fix: Fix nullable field not reported

+++ src/main/kotlin/dev/minuk/opampcommander/domain/models/agent/CommunicationStatus.kt
@@ -1,7 +1,7 @@
 package dev.minuk.opampcommander.domain.models.agent
 
 data class CommunicationStatus(
-    val sequenceNum: Int,
+    val sequenceNum: Long,
 ) {
     companion object {
         fun empty(): CommunicationStatus =
